+++ src/missions_lib/__init__.py
@@ -1,5 +1,7 @@
 # __init__.py
+from .take_message_to import Mission1StateMachine
 from .greet_all import Mission2StateMachine
+from .go_home import Mission4StateMachine